Implement AVG
The window functions now have to specify the type of the segment tree
node and what to do with the query result. For simple window functions,
this can be achieved by using a `using` for the node type and importing
the `IdentityTransformer` into the traits scope. More involved
functions, such as AVG can however specify arbitrary functions.